Adding multiple headers

+++ logstore/src/test/java/io/joshworks/fstore/log/record/DataStreamTest.java
@@ -26,7 +26,7 @@ import static org.junit.Assert.assertEquals;
 
 public class DataStreamTest {
 
-    private static final long MAX_ENTRY_SIZE = 1024 * 1024 * 5L;
+    private static final int MAX_ENTRY_SIZE = 1024 * 1024 * 5;
     private static final double CHCKSUM_PROB = 1;
 
     private File file;
